removed unused files, added path aliases and corrected moved some paths to src
