+++ src/components/layout/Layout.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import Navbar from '@components/Navbar'
+
+const Layout = ({children}) => {
+  return (
+    <>
+        <Navbar />
+        {children}
+    </>
+  )
+}
+
+export default Layout
