+++ src/pages/_app.js
@@ -1,9 +1,10 @@
-import "@/styles/globals.css";
-import { ContextProvider } from "../../contexts/ContextProvider";
+import '@styles/globals.css';
+import { ContextProvider } from "@contexts/ContextProvider";
 
 export default function App({ Component, pageProps }) {
   return (
     <ContextProvider>
       <Component {...pageProps} />
-    </ContextProvider>)
+    </ContextProvider>
+  );
 }
